+++ packages/gatsby-cli/src/reporter/reporters/yurnalist/index.js
@@ -1,7 +1,8 @@
 // @flow
 
 const { createReporter } = require(`yurnalist`)
-const convertHrtime = require(`convert-hrtime`)
+const ProgressBar = require(`progress`)
+const calcElapsedTime = require(`../../../util/calc-elapsed-time`)
 
 const VERBOSE = process.env.gatsby_log_level === `verbose`
 const reporter = createReporter({ emoji: true, verbose: VERBOSE })
@@ -29,36 +30,69 @@ module.exports = {
   info: reporter.info.bind(reporter),
   warn: reporter.warn.bind(reporter),
   log: reporter.log.bind(reporter),
-  /**
-   * Time an activity.
-   * @param {string} name - Name of activity.
-   * @returns {string} The elapsed time of activity.
-   */
-  createActivity(name) {
-    const spinner = reporter.activity()
-    const start = process.hrtime()
-    let status
 
-    const elapsedTime = () => {
-      var elapsed = process.hrtime(start)
-      return `${convertHrtime(elapsed)[`seconds`].toFixed(3)} s`
+  createActivity: activity => {
+    let start
+
+    if (activity.type === `spinner`) {
+      const spinner = reporter.activity()
+      let status
+
+      return {
+        update: newState => {
+          if (newState.startTime) {
+            start = newState.startTime
+            spinner.tick(activity.id)
+          }
+          if (newState.status) {
+            status = newState.status
+            spinner.tick(`${activity.id} — ${newState.status}`)
+          }
+        },
+        done: () => {
+          const str = status
+            ? `${activity.id} — ${calcElapsedTime(start)} — ${status}`
+            : `${activity.id} — ${calcElapsedTime(start)}`
+          reporter.success(str)
+          spinner.end()
+        },
+      }
+    }
+
+    if (activity.type === `progress`) {
+      const bar = new ProgressBar(
+        ` [:bar] :current/:total :elapsed s :percent ${activity.id}`,
+        {
+          total: 0,
+          width: 30,
+          clear: true,
+        }
+      )
+      return {
+        update: newState => {
+          if (newState.startTime) {
+            start = newState.startTime
+          }
+          if (newState.total) {
+            bar.total = newState.total
+          }
+          if (newState.current) {
+            bar.tick()
+          }
+        },
+        done: () => {
+          reporter.success(
+            `${activity.id} — ${bar.curr}/${bar.total} - ${calcElapsedTime(
+              start
+            )} s`
+          )
+        },
+      }
     }
 
     return {
-      start: () => {
-        spinner.tick(name)
-      },
-      setStatus: s => {
-        status = s
-        spinner.tick(`${name} — ${status}`)
-      },
-      end: () => {
-        const str = status
-          ? `${name} — ${elapsedTime()} — ${status}`
-          : `${name} — ${elapsedTime()}`
-        reporter.success(str)
-        spinner.end()
-      },
+      update: () => {},
+      done: () => {},
     }
   },
 }