+++ packages/gatsby-plugin-sharp/src/index.js
@@ -10,7 +10,7 @@ const { scheduleJob } = require(`./scheduler`)
 const { createArgsDigest } = require(`./process-file`)
 const { reportError } = require(`./report-error`)
 const { getPluginOptions, healOptions } = require(`./plugin-options`)
-const { memoizedTraceSVG } = require(`./trace-svg`)
+const { memoizedTraceSVG, notMemoizedtraceSVG } = require(`./trace-svg`)
 
 const imageSizeCache = new Map()
 const getImageSize = file => {
@@ -110,7 +110,8 @@ function queueImageResizing({ file, args = {}, reporter }) {
   const finishedPromise = scheduleJob(
     job,
     boundActionCreators,
-    pluginOptions
+    pluginOptions,
+    reporter
   ).then(() => {
     queue.delete(prefixedSrc)
   })
@@ -144,10 +145,18 @@ async function generateBase64({ file, args, reporter }) {
   })
   let pipeline
   try {
-    pipeline = sharp(file.absolutePath).rotate()
+    pipeline = sharp(file.absolutePath)
+
+    if (!options.rotate) {
+      pipeline.rotate()
+    }
   } catch (err) {
     reportError(`Failed to process image ${file.absolutePath}`, err, reporter)
     return null
+  }
+
+  if (options.trim) {
+    pipeline = pipeline.trim(options.trim)
   }
 
   const forceBase64Format =
@@ -202,40 +211,49 @@ async function generateBase64({ file, args, reporter }) {
   return base64output
 }
 
-const base64CacheKey = ({ file, args }) => `${file.id}${JSON.stringify(args)}`
+const generateCacheKey = ({ file, args }) => `${file.id}${JSON.stringify(args)}`
 
-const memoizedBase64 = _.memoize(generateBase64, base64CacheKey)
+const memoizedBase64 = _.memoize(generateBase64, generateCacheKey)
 
-const cachifiedBase64 = async ({ cache, ...arg }) => {
-  const cacheKey = base64CacheKey(arg)
+const cachifiedProcess = async ({ cache, ...arg }, genKey, processFn) => {
+  const cachedKey = genKey(arg)
+  const cached = await cache.get(cachedKey)
 
-  const cachedBase64 = await cache.get(cacheKey)
-  if (cachedBase64) {
-    return cachedBase64
+  if (cached) {
+    return cached
   }
 
-  const base64output = await generateBase64(arg)
+  const result = await processFn(arg)
+  await cache.set(cachedKey, result)
 
-  await cache.set(cacheKey, base64output)
-
-  return base64output
+  return result
 }
 
 async function base64(arg) {
   if (arg.cache) {
     // Not all tranformer plugins are going to provide cache
-    return await cachifiedBase64(arg)
+    return await cachifiedProcess(arg, generateCacheKey, generateBase64)
   }
 
   return await memoizedBase64(arg)
 }
 
-async function getTracedSVG(options, file) {
+async function traceSVG(args) {
+  if (args.cache) {
+    // Not all tranformer plugins are going to provide cache
+    return await cachifiedProcess(args, generateCacheKey, notMemoizedtraceSVG)
+  }
+  return await memoizedTraceSVG(args)
+}
+
+async function getTracedSVG({ file, options, cache, reporter }) {
   if (options.generateTracedSVG && options.tracedSVG) {
     const tracedSVG = await traceSVG({
-      file,
       args: options.tracedSVG,
       fileArgs: options,
+      file,
+      cache,
+      reporter,
     })
     return tracedSVG
   }
@@ -301,7 +319,7 @@ async function fluid({ file, args = {}, reporter, cache }) {
 
   // Create sizes (in width) for the image if no custom breakpoints are
   // provided. If the max width of the container for the rendered markdown file
-  // is 800px, the sizes would then be: 200, 400, 800, 1200, 1600, 2400.
+  // is 800px, the sizes would then be: 200, 400, 800, 1200, 1600.
   //
   // This is enough sizes to provide close to the optimal image size for every
   // device size / screen resolution while (hopefully) not requiring too much
@@ -316,7 +334,6 @@ async function fluid({ file, args = {}, reporter, cache }) {
     fluidSizes.push(options[fixedDimension] / 2)
     fluidSizes.push(options[fixedDimension] * 1.5)
     fluidSizes.push(options[fixedDimension] * 2)
-    fluidSizes.push(options[fixedDimension] * 3)
   } else {
     options.srcSetBreakpoints.forEach(breakpoint => {
       if (breakpoint < 1) {
@@ -372,6 +389,7 @@ async function fluid({ file, args = {}, reporter, cache }) {
       duotone: options.duotone,
       grayscale: options.grayscale,
       rotate: options.rotate,
+      trim: options.trim,
       toFormat: options.toFormat,
       toFormatBase64: options.toFormatBase64,
       width: base64Width,
@@ -381,7 +399,7 @@ async function fluid({ file, args = {}, reporter, cache }) {
     base64Image = await base64({ file, args: base64Args, reporter, cache })
   }
 
-  const tracedSVG = await getTracedSVG(options, file)
+  const tracedSVG = await getTracedSVG({ options, file, cache, reporter })
 
   // Construct src and srcSet strings.
   const originalImg = _.maxBy(images, image => image.width).src
@@ -436,12 +454,11 @@ async function fixed({ file, args = {}, reporter, cache }) {
   // if no width is passed, we need to resize the image based on the passed height
   const fixedDimension = options.width === undefined ? `height` : `width`
 
-  // Create sizes for different resolutions — we do 1x, 1.5x, 2x, and 3x.
+  // Create sizes for different resolutions — we do 1x, 1.5x, and 2x.
   const sizes = []
   sizes.push(options[fixedDimension])
   sizes.push(options[fixedDimension] * 1.5)
   sizes.push(options[fixedDimension] * 2)
-  sizes.push(options[fixedDimension] * 3)
   const dimensions = getImageSize(file)
 
   const filteredSizes = sizes.filter(size => size <= dimensions[fixedDimension])
@@ -505,7 +522,7 @@ async function fixed({ file, args = {}, reporter, cache }) {
     })
   }
 
-  const tracedSVG = await getTracedSVG(options, file)
+  const tracedSVG = await getTracedSVG({ options, file, reporter, cache })
 
   const fallbackSrc = images[0].src
   const srcSet = images
@@ -520,9 +537,6 @@ async function fixed({ file, args = {}, reporter, cache }) {
           break
         case 2:
           resolution = `2x`
-          break
-        case 3:
-          resolution = `3x`
           break
         default:
       }
@@ -542,10 +556,6 @@ async function fixed({ file, args = {}, reporter, cache }) {
     originalName: originalName,
     tracedSVG,
   }
-}
-
-async function traceSVG(args) {
-  return await memoizedTraceSVG(args)
 }
 
 function toArray(buf) {