+++ packages/gatsby-plugin-sharp/src/trace-svg.js
@@ -16,7 +16,11 @@ exports.notMemoizedPrepareTraceSVGInputFile = async ({
 }) => {
   let pipeline
   try {
-    pipeline = sharp(file.absolutePath).rotate()
+    pipeline = sharp(file.absolutePath)
+
+    if (!options.rotate) {
+      pipeline.rotate()
+    }
   } catch (err) {
     reportError(`Failed to process image ${file.absolutePath}`, err, reporter)
     return
@@ -114,9 +118,13 @@ exports.notMemoizedtraceSVG = async ({ file, args, fileArgs, reporter }) => {
 
     const optionsSVG = _.defaults(args, defaultArgs)
 
+    // `srcset` attribute rejects URIs with literal spaces
+    const encodeSpaces = str => str.replace(/ /gi, `%20`)
+
     return trace(tmpFilePath, optionsSVG)
       .then(optimize)
       .then(svgToMiniDataURI)
+      .then(encodeSpaces)
   } catch (e) {
     throw e
   }