+++ packages/gatsby-source-filesystem/src/__tests__/create-remote-file-node.js
@@ -19,30 +19,31 @@ jest.mock(`got`, () => {
     stream: jest.fn(),
   }
 })
-jest.mock(
-  `progress`,
-  () =>
-    class ProgressBar {
-      static total = 0
-      static tick = jest.fn(() => (ProgressBar.total -= 1))
 
-      total = ProgressBar.total
-      tick = ProgressBar.tick
-    }
-)
+jest.mock(`gatsby-cli/lib/reporter`, () => {
+  return {
+    createProgress: jest.fn(),
+  }
+})
 jest.mock(`../create-file-node`, () => {
   return {
     createFileNode: jest.fn(),
   }
 })
+const reporter = require(`gatsby-cli/lib/reporter`)
+const progressBar = {
+  start: jest.fn(),
+  total: 0,
+  tick: jest.fn(),
+}
+reporter.createProgress.mockImplementation(() => progressBar)
+
 const got = require(`got`)
-const ProgressBar = require(`progress`)
 const createRemoteFileNode = require(`../create-remote-file-node`)
 const { createFileNode } = require(`../create-file-node`)
 
 beforeEach(() => {
-  ProgressBar.total = 0
-  ProgressBar.tick.mockClear()
+  progressBar.tick.mockClear()
 })
 
 describe(`create-remote-file-node`, () => {
@@ -52,6 +53,7 @@ describe(`create-remote-file-node`, () => {
     cache: {},
     createNode: jest.fn(),
     createNodeId: jest.fn(),
+    reporter,
   }
 
   describe(`basic functionality`, () => {
@@ -73,8 +75,8 @@ describe(`create-remote-file-node`, () => {
 
         expect(value).rejects.toMatch(`wrong url: `)
 
-        expect(ProgressBar.total).toBe(0)
-        expect(ProgressBar.tick).not.toHaveBeenCalled()
+        expect(progressBar.total).toBe(0)
+        expect(progressBar.tick).not.toHaveBeenCalled()
       })
     })
   })
@@ -104,6 +106,11 @@ describe(`create-remote-file-node`, () => {
         pipe: jest.fn(() => gotMock),
         on: jest.fn((mockType, mockCallback) => {
           if (mockType === type) {
+            // got throws on 404/500 so we mimic this behaviour
+            if (response.statusCode === 404) {
+              throw new Error(`Response code 404 (Not Found)`)
+            }
+
             mockCallback(response)
           }
 
@@ -136,7 +143,8 @@ describe(`create-remote-file-node`, () => {
     it(`invokes ProgressBar tick`, async () => {
       await setup()
 
-      expect(ProgressBar.tick).toHaveBeenCalledTimes(1)
+      expect(progressBar.total).toBe(1)
+      expect(progressBar.tick).toHaveBeenCalledTimes(1)
     })
 
     describe(`requesting remote image`, () => {
@@ -177,7 +185,7 @@ describe(`create-remote-file-node`, () => {
         )
       })
 
-      it(`passes custom http heades, if defined`, async () => {
+      it(`passes custom http header, if defined`, async () => {
         await setup({
           httpHeaders: {
             Authorization: `Bearer foobar`,
@@ -192,6 +200,19 @@ describe(`create-remote-file-node`, () => {
             }),
           })
         )
+      })
+
+      it(`fails when 404 is given`, async () => {
+        expect.assertions(1)
+        try {
+          await setup({}, `response`, { statusCode: 404 })
+        } catch (err) {
+          expect(err).toEqual(
+            expect.stringContaining(
+              `failed to process https://images.whatever.com/real-image-trust-me`
+            )
+          )
+        }
       })
     })
   })