+++ www/src/components/homepage/homepage-newsletter.js
@@ -16,12 +16,13 @@ import {
   letterSpacings,
   lineHeights,
   fonts,
+  fontWeights,
 } from "../../utils/presets"
 
-const stripedBorderHeight = space[2]
+const stripedBorderHeight = space[1]
 
 const Container = styled(`div`)`
-  border: 1px solid ${colors.ui.light};
+  border: 1px solid ${colors.ui.border.subtle};
   border-radius: ${radii[2]}px;
   display: flex;
   flex-direction: column;
@@ -32,15 +33,15 @@ const Container = styled(`div`)`
 
   :after {
     border-radius: 0 0 ${radii[2]}px ${radii[2]}px;
-    background: ${colors.ui.whisper}
+    background: ${colors.white}
       repeating-linear-gradient(
         135deg,
-        ${colors.lemon},
-        ${colors.lemon} 20px,
+        ${colors.yellow[40]},
+        ${colors.yellow[40]} 20px,
         transparent 20px,
         transparent 40px,
-        ${colors.mint} 40px,
-        ${colors.mint} 60px,
+        ${colors.teal[40]} 40px,
+        ${colors.teal[40]} 60px,
         transparent 60px,
         transparent 80px
       );
@@ -72,7 +73,7 @@ const Name = styled(`h3`)`
   color: ${colors.lilac};
   font-family: ${fonts.header};
   font-size: ${fontSizes[1]};
-  font-weight: normal;
+  font-weight: ${fontWeights[0]};
   letter-spacing: ${letterSpacings.tracked};
   margin: 0;
   text-transform: uppercase;
@@ -81,6 +82,7 @@ const Name = styled(`h3`)`
 const Title = styled(`h1`)`
   color: ${colors.gatsby};
   font-size: ${fontSizes[4]};
+  font-weight: ${fontWeights[1]};
   line-height: ${lineHeights.dense};
   margin: 0;
   margin-top: ${space[1]};