+++ packages/gatsby-plugin-catch-links/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.1.0](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-catch-links/compare/gatsby-plugin-catch-links@2.0.15...gatsby-plugin-catch-links@2.1.0) (2019-06-20)
+
+**Note:** Version bump only for package gatsby-plugin-catch-links
+
+## [2.0.15](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-catch-links/compare/gatsby-plugin-catch-links@2.0.14...gatsby-plugin-catch-links@2.0.15) (2019-05-22)
+
+**Note:** Version bump only for package gatsby-plugin-catch-links
+
 ## [2.0.14](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-catch-links/compare/gatsby-plugin-catch-links@2.0.13...gatsby-plugin-catch-links@2.0.14) (2019-05-14)
 
 ### Bug Fixes